+++ app/uk/gov/hmrc/eventhub/model/MongoEvent.scala
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2021 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.eventhub.model
+
+import org.bson.types.ObjectId
+import play.api.libs.functional.syntax.unlift
+import play.api.libs.functional.syntax._
+import play.api.libs.json._
+import uk.gov.hmrc.mongo.play.json.formats.{MongoFormats, MongoJavatimeFormats}
+
+import java.time.Instant
+
+case class MongoEvent(_id: ObjectId, createdAt: Instant, event: Event)
+
+object MongoEvent {
+  implicit val oif: Format[ObjectId] = MongoFormats.objectIdFormat
+  implicit val instantF: Format[Instant] = MongoJavatimeFormats.instantFormat
+  implicit val fmt: OFormat[MongoEvent] = ((JsPath \ "_id").format[ObjectId]
+    ~ (JsPath \ "createdAt").format[Instant]
+    ~ (JsPath \ "event").format[Event]
+    )(MongoEvent.apply, unlift(MongoEvent.unapply))
+
+  def newMongoEvent(instant: Instant, e: Event): MongoEvent = MongoEvent(ObjectId.get(), instant, e)
+}
